+++ docs/examples/01-csv-transform.php
@@ -1,0 +1,10 @@
+<?php
+
+require_once __DIR__ . "/.init.php";
+
+$chainProcessor = getChainProcessor(__FILE__);
+
+$chainProcessor->process(
+    new ArrayIterator([__DIR__ . "/customers.csv"]),
+    []
+);