+++ src/Oliverde8/Component/PhpEtl/Item/AsyncItemInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Oliverde8\Component\PhpEtl\Item;
+
+interface AsyncItemInterface extends ItemInterface
+{
+    public function isRunning(): bool;
+
+    public function getItem(): ItemInterface;
+}